Make some changes to the original malware scanner branch
