+++ insights/client/apps/malware_detection/__init__.py
@@ -20,41 +20,6 @@ from insights.client.utilities import (
 )
 from insights.util.subproc import call, CalledProcessError
 
-from insights.core.context import HostContext
-from insights.core.dr import SkipComponent
-from insights.specs import Specs
-from insights.core.plugins import datasource
-from insights.core.spec_factory import DatasourceProvider
-
-
-class MalwareDetectionSpecs(Specs):
-    @datasource(HostContext)
-    def run_malware_detection_app(broker):
-        # DatasourceProvider is a generic object for adding content to the archive
-        #   content = the file content to write
-        #   relative_path = filepath to which to write, relative to "data" dir in the archive
-
-        try:
-            insights_config = InsightsConfig(_print_errors=True)
-            insights_config.load_all()
-            if not os.path.isfile(MALWARE_CONFIG_FILE):
-                logger.info("Writing the malware-detection app default configuration to %s", MALWARE_CONFIG_FILE)
-                write_data_to_file(DEFAULT_MALWARE_CONFIG, MALWARE_CONFIG_FILE)
-            try:
-                with open(MALWARE_CONFIG_FILE) as m:
-                    malware_config = yaml.safe_load(m)
-            except Exception as e:
-                logger.error("Error encountered loading the malware-detection app config file %s:\n%s",
-                             MALWARE_CONFIG_FILE, str(e))
-                exit(constants.sig_kill_bad)
-            mdc = MalwareDetectionClient(insights_config, malware_config)
-            scan_results = mdc.run()
-
-            return DatasourceProvider(content=scan_results, relative_path="malware-detection-results.json")
-        except Exception:
-            raise SkipComponent
-
-
 logger = logging.getLogger(__name__)
 MALWARE_UI_URL = 'https://console.redhat.com/insights/malware'
 MIN_YARA_VERSION = "4.1.0"
@@ -143,8 +108,19 @@ cpu_thread_limit: # 2
 
 
 class MalwareDetectionClient:
-    def __init__(self, insights_config, malware_config):
-        self.insights_config = insights_config
+    def __init__(self):
+        self.insights_config = InsightsConfig(_print_errors=True)
+        self.insights_config.load_all()
+        if not os.path.isfile(MALWARE_CONFIG_FILE):
+            logger.info("Writing the malware-detection app default configuration to %s", MALWARE_CONFIG_FILE)
+            write_data_to_file(DEFAULT_MALWARE_CONFIG, MALWARE_CONFIG_FILE)
+        try:
+            with open(MALWARE_CONFIG_FILE) as m:
+                malware_config = yaml.safe_load(m)
+        except Exception as e:
+            logger.error("Error encountered loading the malware-detection app config file %s:\n%s",
+                            MALWARE_CONFIG_FILE, str(e))
+            exit(constants.sig_kill_bad)
         self.malware_config = malware_config
 
         # Early check if the yara binary exists.  No point continuing if it doesn't
