--- insights/tests/datasources/malware_detection_misc/utils.py
@@ -1,21 +1,0 @@
-import os
-from insights.specs.malware_detection import get_parent_dirs
-
-
-def find_misc_dir(misc_dir_name):
-    for path, dirs, files in os.walk(os.getcwd()):
-        if misc_dir_name in dirs:
-            return os.path.join(path, misc_dir_name)
-
-
-def tld(path):
-    """
-    Return the top level directory for 'path' argument
-    """
-    full_path = os.path.abspath(path)
-    if full_path == '/':
-        return full_path
-
-    parent_dirs = []
-    get_parent_dirs(full_path, parent_dirs)
-    return parent_dirs[0]
