+++ insights/tests/datasources/test_malware_detection.py
@@ -10,8 +10,7 @@ try:
 except ImportError:
     from urllib.parse import quote as urlencode  # python 3
 
-from .malware_detection_misc import utils
-from insights.specs.malware_detection import (
+from insights.client.apps.malware_detection import (
     DEFAULT_MALWARE_CONFIG, MIN_YARA_VERSION, MALWARE_CONFIG_FILE,
     MalwareDetectionClient, logger,
     get_toplevel_dirs, get_parent_dirs, remove_child_items, remove_included_excluded_items,
@@ -29,6 +28,26 @@ try:
 except CalledProcessError:
     pass
 
+
+def find_misc_dir(misc_dir_name):
+    for path, dirs, files in os.walk(os.getcwd()):
+        if misc_dir_name in dirs:
+            return os.path.join(path, misc_dir_name)
+
+
+def tld(path):
+    """
+    Return the top level directory for 'path' argument
+    """
+    full_path = os.path.abspath(path)
+    if full_path == '/':
+        return full_path
+
+    parent_dirs = []
+    get_parent_dirs(full_path, parent_dirs)
+    return parent_dirs[0]
+
+
 # Tests using the caplog fixture fail with python 2.6
 IS_PY26 = sys.version_info < (2, 7)
 PY26_SKIP_REASON = "pytest caplog fixture doesn't work in python 2.6"
@@ -38,7 +57,7 @@ IS_QE = '/home/runner' in os.getcwd()  # Is there a better way to detect the QE 
 QE_SKIP_REASON = "Can't scan pytest process in QE Jenkins environment"
 
 # Miscellanous extra files used by the malware tests are in the malware_detection_misc directory
-MISC_DIR = utils.find_misc_dir('malware_detection_misc')
+MISC_DIR = find_misc_dir('malware_detection_misc')
 
 # Location of yara binary used for testing and its libraries
 # Add MISC_DIR to the path so the system automatically finds the yara binary in there
@@ -120,7 +139,7 @@ class TestMalwareDetectionClient:
         yara = YARA_BINARY if not EXISTING_YARA else EXISTING_YARA
         malopts = yaml.safe_load(DEFAULT_MALWARE_CONFIG)
 
-        with patch("insights.specs.malware_detection.MalwareDetectionClient._get_rules", return_value=TEST_RULE_FILE):
+        with patch("insights.client.apps.malware_detection.MalwareDetectionClient._get_rules", return_value=TEST_RULE_FILE):
             mdc = MalwareDetectionClient(None, malopts)
         assert mdc.yara_binary == yara
         assert mdc.rules_file == TEST_RULE_FILE
@@ -147,7 +166,7 @@ class TestMalwareDetectionClient:
         # can't scan the pytest process
         if os.path.exists(MALWARE_CONFIG_FILE):
             mdc.scan_filesystem()
-            assert "Scanning specified files in %s ..." % utils.tld(MALWARE_CONFIG_FILE) in caplog.text
+            assert "Scanning specified files in %s ..." % tld(MALWARE_CONFIG_FILE) in caplog.text
             assert "Matched rule TEST_RedHatInsightsMalwareDetection in file %s" % MALWARE_CONFIG_FILE in caplog.text
 
         mdc.scan_processes()
@@ -164,7 +183,7 @@ class TestMalwareDetectionClient:
         malopts = yaml.safe_load(DEFAULT_MALWARE_CONFIG)
         malopts['test_scan'] = False
 
-        with patch("insights.specs.malware_detection.MalwareDetectionClient._get_rules",
+        with patch("insights.client.apps.malware_detection.MalwareDetectionClient._get_rules",
                    return_value=COMPILED_RULES_FILE):
             mdc = MalwareDetectionClient(None, malopts)
         assert mdc.rules_file == COMPILED_RULES_FILE
@@ -180,7 +199,7 @@ class TestMalwareDetectionClient:
         # TODO: Check the malware client app behaves in a predictable way if the user specifies invalid option values
         pass
 
-    @patch("insights.specs.malware_detection.MalwareDetectionClient._get_rules", return_value=TEST_RULE_FILE)
+    @patch("insights.client.apps.malware_detection.MalwareDetectionClient._get_rules", return_value=TEST_RULE_FILE)
     def test_successful_find_yara_binary(self, rules, caplog):
         logger.setLevel('DEBUG')
         # Find yara using the system search path
@@ -231,8 +250,8 @@ class TestMalwareDetectionClient:
         assert "Error getting the version of the specified yara binary %s" % yara_binary in caplog.text
         assert exc_info.value.code == constants.sig_kill_bad
 
-    @patch("insights.specs.malware_detection.MalwareDetectionClient._get_rules", return_value=TEST_RULE_FILE)
-    @patch("insights.specs.malware_detection.call")
+    @patch("insights.client.apps.malware_detection.MalwareDetectionClient._get_rules", return_value=TEST_RULE_FILE)
+    @patch("insights.client.apps.malware_detection.call")
     def test_find_yara_version(self, call, rules, caplog):
         # Test checking the version of yara
         malopts = get_malware_options()
@@ -250,7 +269,7 @@ class TestMalwareDetectionClient:
         # Valid versions of yara
         for version in ['4.1', '10.0.0']:
             call.return_value = version
-            with patch("insights.specs.malware_detection.MalwareDetectionClient._build_yara_command", return_value="yara"):
+            with patch("insights.client.apps.malware_detection.MalwareDetectionClient._build_yara_command", return_value="yara"):
                 mdc = MalwareDetectionClient(None, malopts)
             assert mdc.yara_binary
 
@@ -262,7 +281,7 @@ class TestMalwareDetectionClient:
         mdc.scan_filesystem()
         assert "Scan only the specified filesystem item: ['%s']" % COMPILED_RULES_FILE in caplog.text
         assert "Using specified rules file: %s" % COMPILED_RULES_FILE in caplog.text
-        assert "Scanning specified files in %s ..." % utils.tld(COMPILED_RULES_FILE) in caplog.text
+        assert "Scanning specified files in %s ..." % tld(COMPILED_RULES_FILE) in caplog.text
         assert "Matched rule MalwareDetectionClientRule in file %s" % COMPILED_RULES_FILE in caplog.text
         assert "Matched rule MiscellaneousStringsRule in file %s" % COMPILED_RULES_FILE in caplog.text
         # Matched 2 rules
@@ -485,7 +504,7 @@ class TestMalwareDetectionClient:
         mdc.scan_fsobjects = [MISC_DIR.replace('/', '//'), MALWARE_TEST_FILE.replace('/', '//')]
         mdc.scan_filesystem()
         # The file paths will have an extra / at the start of them (because os.path.normpath doesn't remove it)
-        assert "Scanning specified files in %s" % utils.tld(MISC_DIR) in caplog.text
+        assert "Scanning specified files in %s" % tld(MISC_DIR) in caplog.text
         rule_match = mdc.host_scan['MalwareDetectionClientRule']
         assert rule_match[0]['source'] == MALWARE_TEST_FILE
         rule_match = mdc.host_scan['MiscellaneousStringsRule']
